+++ Bugsnag/Bugsnag.m
@@ -37,7 +37,7 @@
 
 static BugsnagClient *bsg_g_bugsnag_client = NULL;
 
-BSG_OBJC_DIRECT_MEMBERS
+
 @implementation Bugsnag
 
 + (BugsnagClient *_Nonnull)start {
@@ -123,28 +123,56 @@ BSG_OBJC_DIRECT_MEMBERS
 + (void)notify:(NSException *)exception {
     if ([self bugsnagReadyForInternalCalls]) {
         BSGPreventInlining(@"Prevent");
-        [self.client notifyErrorOrException:exception stackStripDepth:2 block:nil];
+        [self.client notifyErrorOrException:exception stackStripDepth:2 options:nil block:nil];
+    }
+}
+
++ (void)notify:(NSException *)exception options:(BugsnagErrorOptions *)options{
+    if ([self bugsnagReadyForInternalCalls]) {
+        BSGPreventInlining(@"Prevent");
+        [self.client notifyErrorOrException:exception stackStripDepth:2 options:options block:nil];
     }
 }
 
 + (void)notify:(NSException *)exception block:(BugsnagOnErrorBlock)block {
     if ([self bugsnagReadyForInternalCalls]) {
         BSGPreventInlining(@"inlining");
-        [self.client notifyErrorOrException:exception stackStripDepth:2 block:block];
+        [self.client notifyErrorOrException:exception stackStripDepth:2 options:nil block:block];
+    }
+}
+
++ (void)notify:(NSException *)exception options:(BugsnagErrorOptions *)options block:(BugsnagOnErrorBlock)block {
+    if ([self bugsnagReadyForInternalCalls]) {
+        BSGPreventInlining(@"inlining");
+        [self.client notifyErrorOrException:exception stackStripDepth:2 options:options block:block];
     }
 }
 
 + (void)notifyError:(NSError *)error {
     if ([self bugsnagReadyForInternalCalls]) {
         BSGPreventInlining(@"and");
-        [self.client notifyErrorOrException:error stackStripDepth:2 block:nil];
+        [self.client notifyErrorOrException:error stackStripDepth:2 options:nil block:nil];
+    }
+}
+
++ (void)notifyError:(NSError *)error options:(BugsnagErrorOptions *)options{
+    if ([self bugsnagReadyForInternalCalls]) {
+        BSGPreventInlining(@"and");
+        [self.client notifyErrorOrException:error stackStripDepth:2 options:options block:nil];
     }
 }
 
 + (void)notifyError:(NSError *)error block:(BugsnagOnErrorBlock)block {
     if ([self bugsnagReadyForInternalCalls]) {
         BSGPreventInlining(@"outlining");
-        [self.client notifyErrorOrException:error stackStripDepth:2 block:block];
+        [self.client notifyErrorOrException:error stackStripDepth:2 options:nil block:block];
+    }
+}
+
++ (void)notifyError:(NSError *)error options:(BugsnagErrorOptions *)options block:(BugsnagOnErrorBlock)block {
+    if ([self bugsnagReadyForInternalCalls]) {
+        BSGPreventInlining(@"outlining");
+        [self.client notifyErrorOrException:error stackStripDepth:2 options:options block:block];
     }
 }
 
@@ -327,6 +355,20 @@ BSG_OBJC_DIRECT_MEMBERS
 + (NSString *_Nullable)context {
     if ([self bugsnagReadyForInternalCalls]) {
         return self.client.context;
+    }
+    return nil;
+}
+
++ (NSString *_Nullable)setGroupingDiscriminator:(NSString *_Nullable)groupingDiscriminator {
+    if ([self bugsnagReadyForInternalCalls]) {
+        return [self.client setGroupingDiscriminator:groupingDiscriminator];
+    }
+    return nil;
+}
+
++ (NSString *_Nullable)groupingDiscriminator {
+    if ([self bugsnagReadyForInternalCalls]) {
+        return [self.client groupingDiscriminator];
     }
     return nil;
 }