+++ Bugsnag/Delivery/BSGEventUploader.h
@@ -17,12 +17,16 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-BSG_OBJC_DIRECT_MEMBERS
+
 @interface BSGEventUploader : NSObject
 
 - (instancetype)initWithConfiguration:(BugsnagConfiguration *)configuration notifier:(BugsnagNotifier *)notifier;
 
 - (void)storeEvent:(BugsnagEvent *)event;
+
+- (void)storeEventAndSend:(BugsnagEvent *)event;
+
+- (void)storeEventAndFlush:(BugsnagEvent *)event;
 
 - (void)uploadEvent:(BugsnagEvent *)event completionHandler:(nullable void (^)(void))completionHandler;