+++ Bugsnag/include/Bugsnag/Bugsnag.h
@@ -43,6 +43,9 @@
 #import <Bugsnag/BugsnagSession.h>
 #import <Bugsnag/BugsnagStackframe.h>
 #import <Bugsnag/BugsnagThread.h>
+#import <Bugsnag/BugsnagCaptureOptions.h>
+#import <Bugsnag/BugsnagHttpRequest.h>
+#import <Bugsnag/BugsnagHttpResponse.h>
 
 /**
  * Static access to a Bugsnag Client, the easiest way to use Bugsnag in your app.
@@ -141,29 +144,72 @@ BUGSNAG_EXTERN
 + (void)notify:(NSException *_Nonnull)exception;
 
 /**
+ * Send a custom or caught exception to Bugsnag.
+ *
+ * The exception will be sent to Bugsnag in the background allowing your
+ * app to continue running.
+ *
+ * @param exception The exception.
+ * @param options Defines the set of options to control how the event is generated.
+ */
++ (void)notify:(NSException *_Nonnull)exception
+         options:(BugsnagErrorOptions *_Nullable)options;
+
+/**
  *  Send a custom or caught exception to Bugsnag
  *
- *  @param exception The exception
- *  @param block     A block for optionally configuring the error report
+ *  @param exception The exception.
+ *  @param block A block for optionally configuring the error report.
  */
 + (void)notify:(NSException *_Nonnull)exception
          block:(BugsnagOnErrorBlock _Nullable)block;
 
 /**
+ *  Send a custom or caught exception to Bugsnag
+ *
+ *  @param exception The exception.
+ *  @param options Defines the set of options to control how the event is generated.
+ *  @param block A block for optionally configuring the error report.
+ */
++ (void)notify:(NSException *_Nonnull)exception
+         options:(BugsnagErrorOptions *_Nullable)options
+         block:(BugsnagOnErrorBlock _Nullable)block;
+
+/**
  *  Send an error to Bugsnag
  *
- *  @param error The error
+ *  @param error The error.
  */
 + (void)notifyError:(NSError *_Nonnull)error;
 
 /**
  *  Send an error to Bugsnag
  *
- *  @param error The error
- *  @param block A block for optionally configuring the error report
+ *  @param error The error.
+ *  @param options Defines the set of options to control how the event is generated.
+ */
++ (void)notifyError:(NSError *_Nonnull)error
+         options:(BugsnagErrorOptions *_Nullable)options;
+
+/**
+ *  Send an error to Bugsnag
+ *
+ *  @param error The error.
+ *  @param block A block for optionally configuring the error report.
  */
 + (void)notifyError:(NSError *_Nonnull)error
               block:(BugsnagOnErrorBlock _Nullable)block;
+
+/**
+ *  Send an error to Bugsnag
+ *
+ *  @param error The error.
+ *  @param options Defines the set of options to control how the event is generated.
+ *  @param block A block for optionally configuring the error report.
+ */
++ (void)notifyError:(NSError *_Nonnull)error
+         options:(BugsnagErrorOptions *_Nullable)options
+         block:(BugsnagOnErrorBlock _Nullable)block;
 
 // =============================================================================
 // MARK: - Breadcrumbs
@@ -289,6 +335,16 @@ BUGSNAG_EXTERN
  * Retrieves the context - a general summary of what was happening in the application
  */
 + (NSString *_Nullable)context;
+
+/**
+ * Sets the grouping discriminator
+ */
++ (NSString *_Nullable)setGroupingDiscriminator:(NSString *_Nullable)groupingDiscriminator;
+
+/**
+ * Retrieves the grouping discriminator
+ */
++ (NSString *_Nullable)groupingDiscriminator;
 
 // =============================================================================
 // MARK: - User